Комментарии к 4 и 5 лекции

+++ examples/lection04_05/37_CopyAssignmentOperatorForOtherTypes/car.cpp
@@ -1,12 +1,13 @@
 #include "car.h"
 
-Car::Car(const std::string& color, int speed) : 
-    m_color(color),m_speed(speed)
+// Конструктор с параметрами для инициализации цвета и скорости
+Car::Car(const std::string& color, double speed) : 
+    m_color(color), m_speed(speed)
 {
 }
 
+// Деструктор
 Car::~Car()
 {
+    // Деструктор пустой, т.к. нет динамической памяти для освобождения
 }
-
-
