+++ examples/lection04_05/37_CopyAssignmentOperatorForOtherTypes/car.h
@@ -2,22 +2,34 @@
 #define CAR_H
 
 #include <string>
+
 class Car
 {
 public:
+    // Конструктор по умолчанию
     Car() = default;
-    Car(const std::string& color, int speed);
+    
+    // Конструктор с параметрами для инициализации цвета и скорости
+    Car(const std::string& color, double speed);
+    
+    // Деструктор
     ~Car();
     
-    double get_speed()const{
+    // Геттер для получения скорости автомобиля
+    double get_speed() const
+    {
         return m_speed;
     }
-    std::string get_color() const{
+    
+    // Геттер для получения цвета автомобиля
+    std::string get_color() const
+    {
         return m_color;
     }
+
 private:
-    std::string m_color;
-    double m_speed;
+    std::string m_color;  // Цвет автомобиля
+    double m_speed;       // Скорость автомобиля
 };
 
 #endif // CAR_H
